requirements.txt
bottle~=0.12.19
dateparser~=1.1.0
bs4~=0.0.1
beautifulsoup4~=4.10.0
requests~=2.26.0
SQLAlchemy~=1.4.26
DateTime~=4.3
crawler~=0.0.2
grpcio~=1.42.0
pydantic~=1.8.2